fix: remove hardcoded bfloat16 in triton CE backward
The backward kernel was hardcoding .to(tl.bfloat16) for the output store.
Now relies on Triton's implicit dtype conversion to match the output tensor
dtype, which is correct for float16/float32 as well.

+++ src/kernel/triton/cross_entropy.py
@@ -83,7 +83,7 @@ def _cross_entropy_bwd_kernel(
         p = tl.exp(z - max_val) / sum_exp
         dz = tl.where(cols == target, p - 1.0, p) * dloss
 
-        tl.store(DLOGITS_ptr + base + cols, dz.to(tl.bfloat16), mask=mask)
+        tl.store(DLOGITS_ptr + base + cols, dz, mask=mask)
 
 
 def triton_cross_entropy_fwd(logits, targets):
